Skip plot tests if matplotlib is not installed
